refactor(orchestration): move sendGroupToWaitlist to facade layer
Move sendGroupToWaitlist from App.jsx to orchestration layer:
- Create waitlistOrchestrator.js with sendGroupToWaitlistOrchestrated()
- Replace App.jsx implementation with thin wrapper
- Export from orchestration/index.js facade
- Add export smoke test

This is a verbatim move with no logic changes. Function body is
preserved exactly, only closure variables are passed via deps
(~156 lines of waitlist joining logic with validation, mobile
geolocation, and API calls).

Dependencies: isJoiningWaitlist, currentGroup, mobileFlow,
3 setters, backend, getMobileGeolocation, validateGroupCompat,
isPlayerAlreadyPlaying, showAlertMessage, API_CONFIG, Tennis global.
